% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_weights.R
\name{adjust_weights}
\alias{adjust_weights}
\title{Calculate adjusted weights for a subset of the data. See description at http://opendata.cern.ch/record/329.}
\usage{
adjust_weights(
  complete_data,
  subset_data,
  unadjusted_weight_col = "Weight",
  label_col = "Label"
)
}
\arguments{
\item{complete_data}{The complete dataframe. Must contain a weight column (see `unadjusted_weight_col`) and a label column (see `label_col`)}

\item{subset_data}{The subset of `complete_data` for which you want to compute adjusted weights.}

\item{unadjusted_weight_col}{The name of the column containing the unadjusted weight values.}

\item{label_col}{The name of the column containing the label values (signal vs. background). This column must contain "s" for signal and "b" for background.}
}
\value{
A vector of adjusted weights for the observations in `subset_data`
}
\description{
Calculate adjusted weights for a subset of the data. See description at http://opendata.cern.ch/record/329.
}
\examples{
\dontrun{
# Calculate adjusted weights for a subset of data used in the Kaggle challenge:
kaggle_adjust_weights_t <- adjust_weights(complete_data = higgs_data_orig,
subset_data = higgs_data_orig[higgs_data_orig$KaggleSet == "t",],
unadjusted_weight_col = "Weight",
label_col = "Label")
}
}
