% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_median_sig.R
\name{approx_median_sig}
\alias{approx_median_sig}
\title{Calculate the approximate median significance for a given prediction.}
\usage{
approx_median_sig(predictions, labels, weights, b_reg = 10)
}
\arguments{
\item{predictions}{Vector of predicted signal vs. background values. Should contain "s" for signal and "b" for background.}

\item{labels}{Vector of true labels of signal vs. background values. Should contain "s" for signal and "b" for background.}

\item{weights}{Vector of weights for each observation. Note that if you are predicting for a subset of the data, you should be using `adjust_weights` to calculate the adjusted weight values.}

\item{b_reg}{Regularisation parameter as described here: http://opendata.cern.ch/record/329. The default value determined by CERN is 10.}
}
\value{
The value of the approximate median significance (AMS) of the prediction. AMS is used to measure the success of a classifier.
}
\description{
Calculate the approximate median significance for a given prediction.
}
\examples{
\dontrun{
# Test AMS if we predicted all background
predictions <- rep("b", nrow(higgs_data_orig))
labels <- higgs_data_orig$Label
weights <- higgs_data_orig$Weight
approx_mean_sig(predictions, labels, weights) # give us 0
}
}
