% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_discriminant_analysis.R
\name{fisher_discrim}
\alias{fisher_discrim}
\title{Perform Fisher Discriminant Analysis in the binary classification case.}
\usage{
fisher_discrim(data, label_col = "Label", class_pos, class_neg)
}
\arguments{
\item{data}{Dataset containing a column `label_col` with values for the positive and negative classes.}

\item{label_col}{Column of `data` containing the class labels (should have values of `class_pos` and `class_neg`).}

\item{class_pos}{Value of `label_col` for the positive class.}

\item{class_neg}{Value of `label_col` for the negative class.}
}
\value{
The vector result of FDA in the binary classification case.
}
\description{
Perform Fisher Discriminant Analysis in the binary classification case.
}
