% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higgiesmalls-package.R
\docType{package}
\name{higgiesmalls-package}
\alias{higgiesmalls}
\alias{higgiesmalls-package}
\title{higgiesmalls: Useful Functions for Analysing Higgs Boson Data}
\description{
Functions pertaining to the SM1 / SC1 final project analysing data from the Large Hadron Collider.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/babichmorrowc/higgiesmalls}
  \item Report bugs at \url{https://github.com/babichmorrowc/higgiesmalls/issues}
}

}
\keyword{internal}
