% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutual_Information.R
\name{Plot_Mutual_Information}
\alias{Plot_Mutual_Information}
\title{Plot the mutual information of a family of parameters with [Label]}
\usage{
Plot_Mutual_Information(MI_Data)
}
\arguments{
\item{MI_Data}{Dataframe consisting of two columns [data] and [numeric]}
}
\value{
A ggplot of mutual information
}
\description{
Plot the mutual information of a family of parameters with [Label]
}
