% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_plots.R
\name{plots}
\alias{plots}
\title{Produce a density plot for each variable, plotting the variable given b and s respectively}
\usage{
plots(Parameter, data = higgs_data_na)
}
\arguments{
\item{Parameter}{Name of the column in [data] for which the plot will be created.}

\item{data}{Dataframe including column [Parameter]. Must also include column [Label].}
}
\value{
A ggplot object. This is a density plot for each variable, plotting the variable given b and s respectively.
}
\description{
Produce a density plot for each variable, plotting the variable given b and s respectively
}
