% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_plots.R
\name{plots_pairwise}
\alias{plots_pairwise}
\title{Produce the joint distribution of the var|b and var|s respectively}
\usage{
plots_pairwise(P1, data = higgs_data_na)
}
\arguments{
\item{P1}{Name of the column in [data] for which the plot will be created.}

\item{data}{Dataframe including column [P1]. Must also include column [Label] with b and s values.}
}
\value{
A ggplot object. This shows the joint distributions of [P1] given b and s respectively.
}
\description{
Produce the joint distribution of the var|b and var|s respectively
}
