% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_plots.R
\name{plots_product}
\alias{plots_product}
\title{Produce the product distribution of var|b and var|s}
\usage{
plots_product(P1, data = higgs_data_na)
}
\arguments{
\item{P1}{Name of the column in [data] for which the plot will be created.}

\item{data}{Dataframe including column [P1].}
}
\value{
Prints ???
}
\description{
Produce the product distribution of var|b and var|s
}
