% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_discriminant_analysis.R
\name{scatter_ratio}
\alias{scatter_ratio}
\title{Calculate the ratio of between class scatterness to within class scatterness. This is the ratio maximised by Fisher Discriminant Analysis.}
\usage{
scatter_ratio(data, label_col = "Label", class_pos, class_neg, w)
}
\arguments{
\item{data}{Dataset containing a column `label_col` with values for the positive and negative classes.}

\item{label_col}{Column of `data` containing the class labels (should have values of `class_pos` and `class_neg`).}

\item{class_pos}{Value of `label_col` for the positive class.}

\item{class_neg}{Value of `label_col` for the negative class.}

\item{w}{Vector result of performing FDA (can be the output of a call to `fisher_discrim()`)}
}
\value{
A value for the ratio of between class scatterness to within class scatterness.
}
\description{
Calculate the ratio of between class scatterness to within class scatterness. This is the ratio maximised by Fisher Discriminant Analysis.
}
