% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_discriminant_analysis.R
\name{fisher_discrim}
\alias{fisher_discrim}
\title{Perform Fisher Discriminant Analysis in the binary classification case.}
\usage{
fisher_discrim(data, label_col = "Label", class_pos, class_neg)
}
\arguments{
\item{data}{Dataset containing a column [label_col] with values for the positive and negative classes.}

\item{label_col}{Column of [data] containing the class labels (should have values of [class_pos] and [class_neg]).}

\item{class_pos}{Value of [label_col] for the positive class.}

\item{class_neg}{Value of [label_col] for the negative class.}
}
\value{
The vector result of FDA in the binary classification case.
}
\description{
Perform Fisher Discriminant Analysis in the binary classification case.
}
\examples{
X_p <- matrix(rnorm(400) + 3, nrow = 200)
X_n <- matrix(rnorm(400) - 3, nrow = 200)
input_X <- rbind(X_p, X_n)
input_y <- c(rep(1, nrow(X_p)), rep(-1, nrow(X_n)))
input_Xy <- data.frame(cbind(input_X, input_y))
w_vec <- fisher_discrim(input_Xy, label_col = "input_y", class_pos = 1, class_neg = -1)
# Plot showing the direction of the vector:
ggplot2::ggplot() +
ggplot2::geom_point(data = input_Xy, ggplot2::aes(x = V1, y = V2, color = as.factor(input_y))) +
ggplot2::geom_abline(intercept = 0, slope = w_vec[1] / w_vec[2]) +
ggplot2::labs(x = "Dim 1", y = "Dim 2", color = "class")

}
