% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutual_Information.R
\name{Mutual_Information}
\alias{Mutual_Information}
\title{Calculate the mutual information.}
\usage{
Mutual_Information(parameter)
}
\arguments{
\item{parameter}{Name of the column in [higgs_data_na] to calculate mutual information.}
}
\value{
The mutual information of [parameter].
}
\description{
Calculate the mutual information.
}
